new actor for Lith - page has changed
