+++ lithuania/package.json
@@ -3,7 +3,8 @@
 	"version": "0.0.1",
 	"description": "This is a boilerplate of an Apify actor.",
 	"dependencies": {
-		"apify": "^0.22.4"
+		"apify": "^0.22.4",
+		"moment": "^2.29.1"
 	},
 	"devDependencies": {},
 	"scripts": {
